dnfdaemon: move resolve() and do_transaction() to Goal interface
